feat #23 : UserDataModelType 파일 분리
